requirements.txt
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==2.0.12
click==8.1.7
Flask==3.0.3
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.2.0
python-dotenv==0.19.2
requests==2.26.0
urllib3==1.26.20
Werkzeug==3.0.3
